+++ database/migrations/2020_09_11_084026_create_cotizacions_table.php
@@ -29,9 +29,9 @@ class CreateCotizacionsTable extends Migration
             $table->string('condicion')->nullable(); 
             $table->string('nota_de_seguimiento')->nullable();
             $table->string('ruta_print_document')->nullable();
-            $table->float('sub_total', 8, 2)->nullable();
-            $table->float('iva', 8, 2)->nullable();
-            $table->float('total', 8, 2)->nullable();
+            $table->double('sub_total', 99, 2)->nullable();
+            $table->double('iva', 99, 2)->nullable();
+            $table->double('total', 99, 2)->nullable();
             $table->timestamps();
             $table->softDeletes()->nullable();
         });
